alacarte-speechtotext/requirements.txt
firebase-functions==0.1.0
firebase-admin==6.4.0
google-cloud-speech>=2.25.0